refactor: Rename compare-report internals
