+++ packages/cli/src/cmds/compare-report/Reporter.ts
@@ -1,5 +1,5 @@
 import { ChangeReport } from '../compare/ChangeReport';
 
-export default interface Report {
+export default interface Reporter {
   generateReport(changeReport: ChangeReport, baseDir: string): Promise<string>;
 }
